requirements.txt
torch
torchvision
scikit-learn
streamlit
pillow
matplotlib
